(invoice count, pay fileds int in float) && (add invoice filter))

+++ database/migrations/2022_03_01_130930_create_invoice_items_table.php
@@ -23,8 +23,8 @@ class CreateInvoiceItemsTable extends Migration
             $table->unsignedInteger('item_id');
             $table->foreign('item_id')->references('id')->on('items');
 
-            $table->integer('count')->default(0)->unsigned();
-            $table->integer('price')->default(0)->unsigned();
+            $table->bigInteger('count')->default(0)->unsigned();
+            $table->bigInteger('price')->default(0)->unsigned();
 
             $table->string('measurement')->nullable();
